requirements.txt
streamlit
pandas
numpy
pyyaml
matplotlib